Refactor: service layer interface, implementation 분리
