+++ commons/src/main/java/me/ramos/commons/service/QueryTeamService.java
@@ -1,0 +1,11 @@
+package me.ramos.commons.service;
+
+import java.util.List;
+import me.ramos.commons.domain.model.Team;
+
+public interface QueryTeamService {
+
+    Team findByTeamId(Long teamId);
+
+    List<Team> findAll();
+}
