+++ commons/src/main/java/me/ramos/commons/service/impl/TestCommonServiceImpl.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class TestCommonService {
+public class TestCommonServiceImpl {
 
     public String test() {
         return "Hello Common Service";
